Relocate JS tests, for build & coverage
